requirements.txt
django
djangorestframework
markdown
django-filter
cryptography
python-decouple
django-cors-headers
matplotlib
numpy
scipy